code/LlmEvaluation/requirements.txt
openai
anthropic
groq
google-genai
python-dotenv
openpyxl
pandas